fixed functions.py to have integer number of rays for each ring

+++ src/temgymbasic/functions.py
@@ -87,7 +87,7 @@ def circular_beam(r, outer_radius):
     proportion = num_points_kth_ring/np.sum(num_points_kth_ring)
 
     # resolve this proportion to an integer value, and reverse it
-    num_rays_to_each_ring = np.ceil(proportion*remainder_rays)[::-1]
+    num_rays_to_each_ring = np.ceil(proportion*remainder_rays)[::-1].astype(np.int32)
 
     # We need to decide on where to stop adding the remainder of rays to the
     # rest of the rings. We find this point by summing the rays in each ring
